requirements.txt
# 基础依赖
numpy>=1.21.2
pandas>=1.3.3
opencv-python>=4.5.3.56
python-dotenv>=0.19.0
Pillow>=8.3.2
onnxruntime>=1.8.1

# 后端服务依赖
flask>=2.0.1
flask-cors>=3.0.10
flask-sqlalchemy>=2.5.1
flask-jwt-extended>=4.3.1
pymysql>=1.0.2
sqlalchemy>=1.4.23
passlib>=1.7.4
python-jose>=3.3.0

# 开发工具
pytest>=6.2.5
black>=21.8b0
flake8>=3.9.2
